float at int3

+++ vm/gc.c
@@ -12,14 +12,14 @@ vm_value_t vm_gc_new(vm_gc_t* gc, vm_int_t slots) {
 
 vm_value_t vm_gc_get(vm_gc_t* gc, vm_value_t obj, vm_value_t index) {
   vm_value_t val;
-  val.data = vm_value_to_array(obj)->datas[vm_value_to_int(index)];
-  val.type = vm_value_to_array(obj)->types[vm_value_to_int(index)];
+  val.data = vm_value_to_array(obj)->datas[(size_t) vm_value_to_int(index)];
+  val.type = vm_value_to_array(obj)->types[(size_t) vm_value_to_int(index)];
   return val;
 }
 
 void vm_gc_set(vm_gc_t* gc, vm_value_t obj, vm_value_t index, vm_value_t value) {
-  vm_value_to_array(obj)->datas[vm_value_to_int(index)] = value.data;
-  vm_value_to_array(obj)->types[vm_value_to_int(index)] = value.type;
+  vm_value_to_array(obj)->datas[(size_t) vm_value_to_int(index)] = value.data;
+  vm_value_to_array(obj)->types[(size_t) vm_value_to_int(index)] = value.type;
 }
 
 vm_int_t vm_gc_len(vm_gc_t* gc, vm_value_t obj) {
